Side effect as example

+++ src/main/java/functionalJava/Test.java
@@ -15,6 +15,7 @@ public class Test {
     }
 
     public static int div(int a, int b) {
-        return a / b;
+//        return a / b;
+        return (int) (a / (float) b);
     }
 }
